The miner now uses the latest checkpoint block hash for computing the proof of work

+++ src/miner/src/miner.rs
@@ -297,7 +297,7 @@ impl PurpleMiner {
                         .filter(|s| {
                             let proof = Proof {
                                 edge_bits: solver.solutions.edge_bits as u8,
-                                nonce: s.nonce as u64,
+                                nonce: s.nonce as u32,
                                 nonces: s.proof.to_vec(),
                             };
                             proof.to_difficulty() as u64 >= target_difficulty
